exec/stages/disks: refactor into seperate files
The disks.go file is quite large. Refactor into multiple files for
common functionality, partitioning, raid, and filesystems.
