+++ internal/exec/stages/disks/disks.go
@@ -19,10 +19,8 @@
 package disks
 
 import (
-	"errors"
 	"fmt"
 	"os/exec"
-	"strings"
 
 	"github.com/coreos/ignition/internal/config/types"
 	"github.com/coreos/ignition/internal/distro"
@@ -30,16 +28,11 @@ import (
 	"github.com/coreos/ignition/internal/exec/util"
 	"github.com/coreos/ignition/internal/log"
 	"github.com/coreos/ignition/internal/resource"
-	"github.com/coreos/ignition/internal/sgdisk"
 	"github.com/coreos/ignition/internal/systemd"
 )
 
 const (
 	name = "disks"
-)
-
-var (
-	ErrBadFilesystem = errors.New("filesystem is not of the correct type")
 )
 
 func init() {
@@ -169,300 +162,4 @@ func (s stage) waitOnDevicesAndCreateAliases(devs []string, ctxt string) error {
 	}
 
 	return nil
-}
-
-// createPartitions creates the partitions described in config.Storage.Disks.
-func (s stage) createPartitions(config types.Config) error {
-	if len(config.Storage.Disks) == 0 {
-		return nil
-	}
-	s.Logger.PushPrefix("createPartitions")
-	defer s.Logger.PopPrefix()
-
-	devs := []string{}
-	for _, disk := range config.Storage.Disks {
-		devs = append(devs, string(disk.Device))
-	}
-
-	if err := s.waitOnDevicesAndCreateAliases(devs, "disks"); err != nil {
-		return err
-	}
-
-	for _, dev := range config.Storage.Disks {
-		devAlias := util.DeviceAlias(string(dev.Device))
-
-		err := s.Logger.LogOp(func() error {
-			op := sgdisk.Begin(s.Logger, devAlias)
-			if dev.WipeTable {
-				s.Logger.Info("wiping partition table requested on %q", devAlias)
-				op.WipeTable(true)
-			}
-
-			for _, part := range dev.Partitions {
-				op.CreatePartition(part)
-			}
-
-			if err := op.Commit(); err != nil {
-				return fmt.Errorf("commit failure: %v", err)
-			}
-			return nil
-		}, "partitioning %q", devAlias)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// createRaids creates the raid arrays described in config.Storage.Raid.
-func (s stage) createRaids(config types.Config) error {
-	if len(config.Storage.Raid) == 0 {
-		return nil
-	}
-	s.Logger.PushPrefix("createRaids")
-	defer s.Logger.PopPrefix()
-
-	devs := []string{}
-	for _, array := range config.Storage.Raid {
-		for _, dev := range array.Devices {
-			devs = append(devs, string(dev))
-		}
-	}
-
-	if err := s.waitOnDevicesAndCreateAliases(devs, "raids"); err != nil {
-		return err
-	}
-
-	for _, md := range config.Storage.Raid {
-		args := []string{
-			"--create", md.Name,
-			"--force",
-			"--run",
-			"--homehost", "any",
-			"--level", md.Level,
-			"--raid-devices", fmt.Sprintf("%d", len(md.Devices)-md.Spares),
-		}
-
-		if md.Spares > 0 {
-			args = append(args, "--spare-devices", fmt.Sprintf("%d", md.Spares))
-		}
-
-		for _, o := range md.Options {
-			args = append(args, string(o))
-		}
-
-		for _, dev := range md.Devices {
-			args = append(args, util.DeviceAlias(string(dev)))
-		}
-
-		if _, err := s.Logger.LogCmd(
-			exec.Command(distro.MdadmCmd(), args...),
-			"creating %q", md.Name,
-		); err != nil {
-			return fmt.Errorf("mdadm failed: %v", err)
-		}
-	}
-
-	return nil
-}
-
-// createFilesystems creates the filesystems described in config.Storage.Filesystems.
-func (s stage) createFilesystems(config types.Config) error {
-	fss := make([]types.Mount, 0, len(config.Storage.Filesystems))
-	for _, fs := range config.Storage.Filesystems {
-		if fs.Mount != nil {
-			fss = append(fss, *fs.Mount)
-		}
-	}
-
-	if len(fss) == 0 {
-		return nil
-	}
-	s.Logger.PushPrefix("createFilesystems")
-	defer s.Logger.PopPrefix()
-
-	devs := []string{}
-	for _, fs := range fss {
-		devs = append(devs, string(fs.Device))
-	}
-
-	if err := s.waitOnDevicesAndCreateAliases(devs, "filesystems"); err != nil {
-		return err
-	}
-
-	for _, fs := range fss {
-		if err := s.createFilesystem(fs); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (s stage) createFilesystem(fs types.Mount) error {
-	info, err := s.readFilesystemInfo(fs)
-	if err != nil {
-		return err
-	}
-
-	if fs.Create != nil {
-		// If we are using 2.0.0 semantics...
-
-		if !fs.Create.Force && info.format != "" {
-			s.Logger.Err("filesystem detected at %q (found %s) and force was not requested", fs.Device, info.format)
-			return ErrBadFilesystem
-		}
-	} else if !fs.WipeFilesystem {
-		// If the filesystem isn't forcefully being created, then we need
-		// to check if it is of the correct type or that no filesystem exists.
-
-		if info.format == fs.Format &&
-			(fs.Label == nil || info.label == *fs.Label) &&
-			(fs.UUID == nil || canonicalizeFilesystemUUID(info.format, info.uuid) == canonicalizeFilesystemUUID(fs.Format, *fs.UUID)) {
-			s.Logger.Info("filesystem at %q is already correctly formatted. Skipping mkfs...", fs.Device)
-			return nil
-		} else if info.format != "" {
-			s.Logger.Err("filesystem at %q is not of the correct type, label, or UUID (found %s, %q, %s) and a filesystem wipe was not requested", fs.Device, info.format, info.label, info.uuid)
-			return ErrBadFilesystem
-		}
-	}
-
-	mkfs := ""
-	var args []string
-	if fs.Create == nil {
-		args = translateMountOptionSliceToStringSlice(fs.Options)
-	} else {
-		args = translateCreateOptionSliceToStringSlice(fs.Create.Options)
-	}
-	switch fs.Format {
-	case "btrfs":
-		mkfs = distro.BtrfsMkfsCmd()
-		args = append(args, "--force")
-		if fs.UUID != nil {
-			args = append(args, []string{"-U", canonicalizeFilesystemUUID(fs.Format, *fs.UUID)}...)
-		}
-		if fs.Label != nil {
-			args = append(args, []string{"-L", *fs.Label}...)
-		}
-	case "ext4":
-		mkfs = distro.Ext4MkfsCmd()
-		args = append(args, "-F")
-		if fs.UUID != nil {
-			args = append(args, []string{"-U", canonicalizeFilesystemUUID(fs.Format, *fs.UUID)}...)
-		}
-		if fs.Label != nil {
-			args = append(args, []string{"-L", *fs.Label}...)
-		}
-	case "xfs":
-		mkfs = distro.XfsMkfsCmd()
-		args = append(args, "-f")
-		if fs.UUID != nil {
-			args = append(args, []string{"-m", "uuid=" + canonicalizeFilesystemUUID(fs.Format, *fs.UUID)}...)
-		}
-		if fs.Label != nil {
-			args = append(args, []string{"-L", *fs.Label}...)
-		}
-	case "swap":
-		mkfs = distro.SwapMkfsCmd()
-		args = append(args, "-f")
-		if fs.UUID != nil {
-			args = append(args, []string{"-U", canonicalizeFilesystemUUID(fs.Format, *fs.UUID)}...)
-		}
-		if fs.Label != nil {
-			args = append(args, []string{"-L", *fs.Label}...)
-		}
-	case "vfat":
-		mkfs = distro.VfatMkfsCmd()
-		// There is no force flag for mkfs.vfat, it always destroys any data on
-		// the device at which it is pointed.
-		if fs.UUID != nil {
-			args = append(args, []string{"-i", canonicalizeFilesystemUUID(fs.Format, *fs.UUID)}...)
-		}
-		if fs.Label != nil {
-			args = append(args, []string{"-n", *fs.Label}...)
-		}
-	default:
-		return fmt.Errorf("unsupported filesystem format: %q", fs.Format)
-	}
-
-	devAlias := util.DeviceAlias(string(fs.Device))
-	args = append(args, devAlias)
-	if _, err := s.Logger.LogCmd(
-		exec.Command(mkfs, args...),
-		"creating %q filesystem on %q",
-		fs.Format, devAlias,
-	); err != nil {
-		return fmt.Errorf("mkfs failed: %v", err)
-	}
-
-	return nil
-}
-
-// golang--
-func translateMountOptionSliceToStringSlice(opts []types.MountOption) []string {
-	newOpts := make([]string, len(opts))
-	for i, o := range opts {
-		newOpts[i] = string(o)
-	}
-	return newOpts
-}
-
-// golang--
-func translateCreateOptionSliceToStringSlice(opts []types.CreateOption) []string {
-	newOpts := make([]string, len(opts))
-	for i, o := range opts {
-		newOpts[i] = string(o)
-	}
-	return newOpts
-}
-
-type filesystemInfo struct {
-	format string
-	uuid   string
-	label  string
-}
-
-func (s stage) readFilesystemInfo(fs types.Mount) (filesystemInfo, error) {
-	res := filesystemInfo{}
-	err := s.Logger.LogOp(
-		func() error {
-			var err error
-			res.format, err = util.FilesystemType(fs.Device)
-			if err != nil {
-				return err
-			}
-			res.uuid, err = util.FilesystemUUID(fs.Device)
-			if err != nil {
-				return err
-			}
-			res.label, err = util.FilesystemLabel(fs.Device)
-			if err != nil {
-				return err
-			}
-			s.Logger.Info("found %s filesystem at %q with uuid %q and label %q", res.format, fs.Device, res.uuid, res.label)
-			return nil
-		},
-		"determining filesystem type of %q", fs.Device,
-	)
-
-	return res, err
-}
-
-// canonicalizeFilesystemUUID does the minimum amount of canonicalization
-// required to make two valid equivalent UUIDs compare equal, but doesn't
-// attempt to fully validate the UUID.
-func canonicalizeFilesystemUUID(format, uuid string) string {
-	uuid = strings.ToLower(uuid)
-	if format == "vfat" {
-		// FAT uses a 32-bit volume ID instead of a UUID. blkid
-		// (and the rest of the world) formats it as A1B2-C3D4, but
-		// mkfs.fat doesn't permit the dash, so strip it. Older
-		// versions of Ignition would fail if the config included
-		// the dash, so we need to support omitting it.
-		if len(uuid) >= 5 && uuid[4] == '-' {
-			uuid = uuid[0:4] + uuid[5:]
-		}
-	}
-	return uuid
 }
